Bug 861039 - Update ANGLE to r2042. - r=bjacob
UltraBlame original commit: e9b37726c0202af23be989e63104ef21132681fd

+++ gfx/angle/src/compiler/preprocessor/Preprocessor.h
@@ -7,6 +7,8 @@
 #ifndef COMPILER_PREPROCESSOR_PREPROCESSOR_H_
 #define COMPILER_PREPROCESSOR_PREPROCESSOR_H_
 
+#include <stddef.h>
+
 #include "pp_utils.h"
 
 namespace pp
@@ -32,7 +34,7 @@ class Preprocessor
     
     
     
-    bool init(int count, const char* const string[], const int length[]);
+    bool init(size_t count, const char* const string[], const int length[]);
     
     void predefineMacro(const char* name, int value);
